moved selection functions to jsmd-selection.js
